We now read all the Materials in the shader code as an array.
Also some syntax fixes.

+++ Source/Engine/World Manager/Scene Manager/Lights/ILight.h
@@ -6,7 +6,7 @@
 
 namespace Engine {
 
-	enum class LightType : uint32_t {
+	enum class LightType : uint16_t {
 		Directional,
 		Point,
 		Spot
